+++ commands/errors.go
@@ -57,6 +57,7 @@ var (
 	ErrMustBeEmpty                                     = errors.New("must be empty")
 	ErrMustBeGTEClampLowerBound                        = errors.New("must be greater than or equal to clamp lower bound")
 	ErrOneTimeTriggerAllowedMax                        = errors.New("maximum one time trigger allowed")
+	ErrMustBeGreaterThanEnactmentTimestamp             = errors.New("must be greater than proposal_submission.terms.enactment_timestamp")
 )
 
 type Errors map[string][]error